Completed automatic channel assign implementation

+++ xbee/trunk/libxbee/xbee-conn.h
@@ -117,6 +117,6 @@ void xbee_conn_command_test ( XbeeConn * conn, gchar *data);
 void xbee_conn_register_callbacks (XbeeConn *conn, xb_conn_callbacks_t *callbacks);
 
 /* Sets the channel of the client associated with the XbeeConn */
-void xbee_conn_set_channel ( XbeeConn *conn, uint8_t channel );
+void xbee_conn_set_channel ( XbeeConn *conn, int16_t channel );
 
 #endif	/* __XBEE_CONN_H */
